[refactor] Category -> Challenge Category 이동
