+++ Routinus/Routinus/Domain/Entity/TodayRoutine.swift
@@ -11,12 +11,12 @@ import RoutinusDatabase
 
 struct TodayRoutine {
     let challengeID: String
-    let category: Category
+    let category: Challenge.Category
     let title: String
     let authCount: Int
     let totalCount: Int
     
-    init(challengeID: String, category: Category, title: String, authCount: Int, totalCount: Int) {
+    init(challengeID: String, category: Challenge.Category, title: String, authCount: Int, totalCount: Int) {
         self.challengeID = challengeID
         self.category = category
         self.title = title
@@ -26,7 +26,7 @@ struct TodayRoutine {
 
     init(todayRoutineDTO: TodayRoutineDTO) {
         self.challengeID = todayRoutineDTO.challengeID
-        self.category = Category.category(by: todayRoutineDTO.categoryID)
+        self.category = Challenge.Category.category(by: todayRoutineDTO.categoryID)
         self.title = todayRoutineDTO.title
         self.authCount = todayRoutineDTO.authCount
         self.totalCount = Date.days(from: todayRoutineDTO.endDate, to: todayRoutineDTO.joinDate)
