+++ Routinus/Routinus/Presentation/Challenge/ChallengeViewModel.swift
@@ -12,7 +12,7 @@ protocol ChallengeViewModelInput {
     func didTappedSearchButton()
     func didTappedSeeAllButton()
     func didTappedRecommendChallenge(index: Int)
-    func didTappedCategoryButton(category: Category)
+    func didTappedCategoryButton(category: Challenge.Category)
 }
 
 protocol ChallengeViewModelOutput {
@@ -20,7 +20,7 @@ protocol ChallengeViewModelOutput {
 
     var showChallengeSearchSignal: PassthroughSubject<Void, Never> { get }
     var showChallengeDetailSignal: PassthroughSubject<String, Never> { get }
-    var showChallengeCategorySignal: PassthroughSubject<Category, Never> { get }
+    var showChallengeCategorySignal: PassthroughSubject<Challenge.Category, Never> { get }
 }
 
 protocol ChallengeViewModelIO: ChallengeViewModelInput, ChallengeViewModelOutput { }
@@ -30,7 +30,7 @@ class ChallengeViewModel: ChallengeViewModelIO {
 
     var showChallengeSearchSignal = PassthroughSubject<Void, Never>()
     var showChallengeDetailSignal = PassthroughSubject<String, Never>()
-    var showChallengeCategorySignal = PassthroughSubject<Category, Never>()
+    var showChallengeCategorySignal = PassthroughSubject<Challenge.Category, Never>()
     
     let usecase: ChallengeFetchableUsecase
     var cancellables = Set<AnyCancellable>()
@@ -55,7 +55,7 @@ extension ChallengeViewModel {
         showChallengeDetailSignal.send(challengeID)
     }
 
-    func didTappedCategoryButton(category: Category) {
+    func didTappedCategoryButton(category: Challenge.Category) {
         showChallengeCategorySignal.send(category)
     }
 }
