+++ Routinus/Routinus/Presentation/Challenge/cell/CategoryButtonTapGesture.swift
@@ -8,9 +8,9 @@
 import UIKit
 
 class CategoryButtonTapGesture: UITapGestureRecognizer {
-    private (set) var category: Category?
+    private (set) var category: Challenge.Category?
     
-    func configureCategory(category: Category) {
+    func configureCategory(category: Challenge.Category) {
         self.category = category
     }
 }
